Added bayesx.construct.default() which essentially just throws an error and recommends to used sx(). The function was added to the NAMESPACE, also adding some more comments.
Added a vector available.terms within sx() so that no other types of
terms can be constructed with sx(). Please check if the list is complete
and correct.

Reworked the parts of the R2BayesX.Rnw paper that involve the function s().
It is now not mentioned before Section 4.2 and just briefly mentioned
twice after that section. All further discussion of s() was deferred to
a new Appendix C that requires additional input and rewriting from NU.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/bayesr@331 edb9625f-4e0d-4859-8d74-9fd3b1da38cb